+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added: ``include`` option that will add ``include`` to all loaders definition.
 - Fixed: regex for extension where not having the proper ``\`` before the dot.
 
 # 0.1.0 - 2016-03-30